requirements.txt
Flask
pygal
IPy
geoip2
python-dateutil
gunicorn
